refactor: consolidate course renames into single file in apps/antalmanac/src/lib
Per feedback:
- Remove packages/types/src/courseRenames.ts and courseRenameUtils.ts
- Remove their re-exports from packages/types/src/index.ts
- Add apps/antalmanac/src/lib/courseRenames.ts with everything in one place

Data structure change: nested Record<department, Record<courseNumber, entry>>
instead of a flat map keyed on composite strings. The previously field also
stores { department, courseNumber } directly, eliminating toCourseId /
fromCourseId encode-decode entirely. Cycle guard uses a simple loop-count
bound instead of a visited Set over encoded strings.

No logic changes — getAllCourseIdentifiers and mergeAggregateGrades work
the same way; routers updated to call getAllCourseIdentifiers(dept, num).
