+++ apps/antalmanac/src/backend/routers/enrollHist.ts
@@ -1,6 +1,6 @@
 import { aapiClient, aapiProcedure } from '$src/backend/lib/aapi';
+import { getAllCourseIdentifiers } from '$src/lib/courseRenames';
 import { WebsocSectionTypeSchema } from '@packages/antalmanac-types';
-import { getAllCourseIdentifiers } from '@packages/antalmanac-types';
 import type { EnrollmentHistoryEntry } from '@packages/anteater-api/types';
 import { z } from 'zod';
 
@@ -17,7 +17,7 @@ const enrollHistRouter = router({
         )
         .query(async ({ input }): Promise<EnrollmentHistoryEntry[]> => {
             const { department, courseNumber, sectionType } = input;
-            const identifiers = getAllCourseIdentifiers({ department, courseNumber });
+            const identifiers = getAllCourseIdentifiers(department, courseNumber);
 
             // Fan out across all predecessor course IDs in parallel.
             const results = await Promise.all(
@@ -31,9 +31,8 @@ const enrollHistRouter = router({
             );
 
             // Enrollment history entries are disjoint across rename boundaries
-            // (data for each academic year lives under whichever course ID was
-            // active at the time), so concatenation is safe — no de-duplication
-            // is needed.
+            // (each year's data lives under whichever course ID was active then),
+            // so concatenation is safe with no de-duplication needed.
             // FIXME: remove this filter once the API stops returning entries with empty date arrays
             return results.flat().filter((x: EnrollmentHistoryEntry) => x.dates.length);
         }),
